Autocorrect Style/IfUnlessModifier

+++ publify_textfilter_code/spec/dummy/db/migrate/20160716093730_add_blog_ids.publify_core_engine.rb
@@ -23,9 +23,7 @@ class AddBlogIds < ActiveRecord::Migration[4.2]
   def down
     if adapter_name == 'PostgreSQL'
       indexes(:contents).each do |index|
-        if index.name =~ /blog_id/
-          remove_index(:contents, name: index.name)
-        end
+        remove_index(:contents, name: index.name) if index.name =~ /blog_id/
       end
     else
       remove_index :contents, :blog_id rescue nil
